+++ .github/workflows/pull_request_badge.js
@@ -1,6 +1,7 @@
 const fs = require("fs");
 const yaml = require("js-yaml");
-const axios = require("axios");
+const github = require("@actions/github");
+const pullRequest = github.context.payload.pull_request;
 
 const configPath = ".github/pr-badge.yml";
 const issuePrefixRegex = /^(\w+-\d+)/i;
@@ -11,23 +12,6 @@ function getIssuePrefix(title) {
 }
 
 async function getPullRequestInfo() {
-  console.log("no");
-  const { client_payload } = require(process.env.GITHUB_EVENT_PATH);
-  console.log("Etest3");
-  console.log(process.env);
-  console.log(process.env.TERA);
-
-  return 0;
-  const response = await axios.get(process.env.GITHUB_EVENT_PATH, {
-    headers: {
-      Authorization: `Bearer ${process.env.GITHUB_TOKEN}`,
-      Accept: "application/vnd.github.v3+json",
-    },
-  });
-
-  console.log(response);
-
-  const pullRequest = response.data;
   const title = pullRequest.title;
   const branchName = pullRequest.head.ref;
   const issuePrefix = getIssuePrefix(title) || getIssuePrefix(branchName);
@@ -51,7 +35,11 @@ function evaluateCondition(condition, pullRequestInfo) {
     title: pullRequestInfo.title,
     branchName: pullRequestInfo.branchName,
     size: pullRequestInfo.size,
+    title: pullRequestInfo.title,
   };
+
+  preview_available = true;
+  size = vars.size;
 
   const code =
     Object.keys(vars)
@@ -61,7 +49,27 @@ function evaluateCondition(condition, pullRequestInfo) {
   return eval(code);
 }
 
-function createBadgeMarkdown(badge) {
+function createBadgeMarkdown(badge, pullRequestInfo) {
+  if (badge.name === "Preview") {
+    badge.link =
+      "http://" + pullRequestInfo.branchName + ".front-a7u.pages.dev/";
+  }
+  if (badge.name === "Redmine") {
+    // an reg ex that take first numbers from Xstring until a "-"
+    const redmineId = pullRequestInfo.title.match(/\d+/)[0];
+    badge.link = "https://redmine.bredband2.se/" + redmineId;
+    badge.image = `https://img.shields.io/badge/Redmine-${redmineId}-red`;
+  }
+  if (badge.name === "Test") {
+    // a reg exthat check for " # Test" and an paragraph after it in the body of the pull request
+    const test = pullRequestInfo.pullRequest.body.match(/# Testguide[\s\S]/);
+    if (test) {
+      badge.image = `https://img.shields.io/badge/Test_Plan-pass-green`;
+    } else {
+      badge.image = `https://img.shields.io/badge/Missing-test_plan-red`;
+    }
+  }
+
   const image = badge.image;
   const link = badge.link || "";
 
@@ -69,16 +77,13 @@ function createBadgeMarkdown(badge) {
 }
 
 async function run() {
-  const { client_payload } = require(process.env.GITHUB_EVENT_PATH);
-  console.log(client_payload);
-  return 0;
-
   const pullRequestInfo = await getPullRequestInfo();
-  const config = yaml.safeLoad(fs.readFileSync(configPath, "utf8"));
-
+  const config = yaml.load(fs.readFileSync(configPath, "utf8"));
   const badges = config.badges
     .filter((badge) => evaluateCondition(badge.condition, pullRequestInfo))
-    .map(createBadgeMarkdown);
+    .map(function (badge) {
+      return createBadgeMarkdown(badge, pullRequestInfo);
+    });
 
   if (badges.length === 0) {
     console.log("No badges to display");
@@ -86,20 +91,18 @@ async function run() {
   }
 
   const body = pullRequestInfo.pullRequest.body || "";
-  const newBody = `${badges.join("\n")}\n\n${body}`;
+  //remove all line starting with [![
+  const bodyWithoutBadges2 = body.replace(/\[!\[.*\]\(.*\)\]\(.*\)/g, "");
+  const newBody = `${badges.join(" ")}\n\n${bodyWithoutBadges2}`;
 
-  await axios.patch(
-    pullRequestInfo.pullRequest.url,
-    {
-      body: newBody,
-    },
-    {
-      headers: {
-        Authorization: `Bearer ${process.env.GITHUB_TOKEN}`,
-        Accept: "application/vnd.github.v3+json",
-      },
-    }
-  );
+  const octokit = github.getOctokit(process.env.GITHUB_TOKEN);
+
+  const { data: pullRequest2 } = await octokit.rest.pulls.update({
+    owner: "ehsanpo",
+    repo: "nextjs-portfolio",
+    pull_number: pullRequest.number,
+    body: newBody,
+  });
 
   console.log(
     `Added ${badges.length} badge(s) to Pull Request #${pullRequestInfo.pullRequest.number}`
