+++ action.yml
@@ -6,4 +6,4 @@ inputs:
     required: true
 runs:
   using: "node16"
-  main: "./pull_request_badge.js"
+  main: "./.github/workflows/pull_request_badge.js"
